+++ Content.Shared/_Sunrise/LockableEquipment/EquipmentContainerSystem.cs
@@ -30,9 +30,11 @@ public sealed class EquipmentContainerSystem : EntitySystem
     public override void Initialize()
     {
         SubscribeLocalEvent<EquipmentContainerComponent, InteractUsingEvent>(OnInteractUsing);
+        SubscribeLocalEvent<EquipmentContainerComponent, GetVerbsEvent<InteractionVerb>>(OnGetVerbs);
         SubscribeLocalEvent<EquipmentContainerComponent, EquipmentDoAfterEvent>(OnDoAfter);
         SubscribeLocalEvent<EquipmentContainerComponent, EntInsertedIntoContainerMessage>(OnContainerInserted);
         SubscribeLocalEvent<EquipmentContainerComponent, EntRemovedFromContainerMessage>(OnContainerRemoved);
+
         SubscribeLocalEvent<EquipmentContainerComponent, EquipmentContainerUseHeldKeyVerbEvent>(OnUseHeldKeyVerb);
         SubscribeLocalEvent<EquipmentContainerComponent, EquipmentContainerBreakWithHeldToolVerbEvent>(OnBreakWithHeldToolVerb);
         SubscribeLocalEvent<EquipmentContainerComponent, EquipmentContainerRemoveVerbEvent>(OnRemoveVerb);
@@ -71,7 +73,6 @@ public sealed class EquipmentContainerSystem : EntitySystem
                 args.Handled = true;
                 return;
             }
-
         }
 
         if (!TryComp(args.Used, out LockableEquipmentComponent? device))
@@ -81,88 +82,69 @@ public sealed class EquipmentContainerSystem : EntitySystem
         args.Handled = TryAttachDevice(ent, args.User, args.Used, device, container);
     }
 
-    /// <summary>
-    /// Attempts to remove the currently installed device from the target.
-    /// </summary>
-    public void TryRemove(EntityUid target, EntityUid user)
+    private void OnGetVerbs(Entity<EquipmentContainerComponent> ent, ref GetVerbsEvent<InteractionVerb> args)
     {
-        if (user == EntityUid.Invalid || Deleted(user))
+        if (!args.CanInteract || !args.CanAccess)
             return;
 
-        if (!TryComp(target, out EquipmentContainerComponent? comp))
+        if (!TryGetEquipment(ent.Owner, ent.Comp, out var device))
             return;
 
-        EnsureComp<DoAfterComponent>(user);
-
-        var container = _container.EnsureContainer<ContainerSlot>(target, comp.ContainerId);
-        if (!CanRemove(target, user, container))
+        if (!TryComp(device, out LockableEquipmentComponent? comp))
             return;
 
-        var doAfter = new DoAfterArgs(
-            EntityManager,
-            user,
-            comp.DetachDoAfter,
-            new EquipmentDoAfterEvent(EquipmentActionType.Detach),
-            target,
-            target: target)
+        if (!CanAccess(ent.Owner, comp.Layer, comp))
+            return;
+
+        var name = MetaData(device).EntityName;
+        var addedKeyVerb = false;
+        var addedBreakVerb = false;
+
+        foreach (var hand in _hands.EnumerateHands(args.User))
         {
-            BreakOnMove = true,
-            BreakOnDamage = true,
-            NeedHand = true,
-            BreakOnHandChange = true,
-            DuplicateCondition = DuplicateConditions.SameTarget | DuplicateConditions.SameEvent
-        };
+            if (!_hands.TryGetHeldItem(args.User, hand, out var held))
+                continue;
 
-        _doAfter.TryStartDoAfter(doAfter);
-    }
-
-    /// <summary>
-    /// Attempts to start the attach flow for a device held by the user.
-    /// </summary>
-    public bool TryAttachDevice(Entity<EquipmentContainerComponent> ent, EntityUid user, EntityUid deviceUid, LockableEquipmentComponent device, BaseContainer? container = null)
-    {
-        if (user == EntityUid.Invalid || Deleted(user))
-            return false;
-
-        container ??= _container.EnsureContainer<ContainerSlot>(ent.Owner, ent.Comp.ContainerId);
-
-        if (!CanAttachDevice(ent.Owner, device, container))
-        {
-            if (!CanAccess(ent.Owner, device.Layer, device))
+            if (!addedKeyVerb && !comp.Broken && HasComp<KeyComponent>(held.Value))
             {
-                _popup.PopupClient(
-                    Loc.GetString("lockable-equipment-blocked"),
-                    user);
-            }
-            else if (FindDevice(container) != null)
-            {
-                _popup.PopupClient(
-                    Loc.GetString("lockable-equipment-already"),
-                    user);
+                addedKeyVerb = true;
+
+                args.Verbs.Add(new InteractionVerb
+                {
+                    Text = comp.Locked
+                        ? Loc.GetString("lockable-equipment-verb-unlock", ("name", name))
+                        : Loc.GetString("lockable-equipment-verb-lock", ("name", name)),
+                    EventTarget = ent,
+                    ExecutionEventArgs = new EquipmentContainerUseHeldKeyVerbEvent(args.User)
+                });
             }
 
-            return false;
+            if (!addedBreakVerb && comp.Locked && _lockable.CanBreakWithTool(device, held.Value, comp))
+            {
+                var breakText = GetBreakVerbText(name, comp.Mode);
+                if (breakText == null)
+                    continue;
+
+                addedBreakVerb = true;
+
+                args.Verbs.Add(new InteractionVerb
+                {
+                    Text = breakText,
+                    EventTarget = ent,
+                    ExecutionEventArgs = new EquipmentContainerBreakWithHeldToolVerbEvent(args.User)
+                });
+            }
         }
 
-        EnsureComp<DoAfterComponent>(user);
-
-        var doAfter = new DoAfterArgs(
-            EntityManager,
-            user,
-            ent.Comp.AttachDoAfter,
-            new EquipmentDoAfterEvent(EquipmentActionType.Attach),
-            ent.Owner,
-            target: ent.Owner,
-            used: deviceUid)
+        if (!comp.Locked)
         {
-            BreakOnMove = true,
-            BreakOnDamage = true,
-            NeedHand = true,
-            BreakOnHandChange = true,
-            DuplicateCondition = DuplicateConditions.SameTarget | DuplicateConditions.SameEvent
-        };
-
-        return _doAfter.TryStartDoAfter(doAfter);
+            args.Verbs.Add(new InteractionVerb
+            {
+                Text = Loc.GetString("lockable-equipment-verb-remove", ("name", name)),
+                EventTarget = ent,
+                ExecutionEventArgs = new EquipmentContainerRemoveVerbEvent(args.User)
+            });
+        }
     }
 
     private void OnDoAfter(Entity<EquipmentContainerComponent> ent, ref EquipmentDoAfterEvent args)
@@ -225,14 +207,148 @@ public sealed class EquipmentContainerSystem : EntitySystem
         args.Handled = true;
     }
 
-    //Helpers
+    private void OnContainerInserted(Entity<EquipmentContainerComponent> ent, ref EntInsertedIntoContainerMessage args)
+    {
+        if (_timing.ApplyingState)
+            return;
+
+        if (args.Container.ID != ent.Comp.ContainerId)
+            return;
+
+        if (TryComp(args.Entity, out LockableEquipmentComponent? device))
+            UpdateAppearance(ent.Owner, device);
+    }
+
+    private void OnContainerRemoved(Entity<EquipmentContainerComponent> ent, ref EntRemovedFromContainerMessage args)
+    {
+        if (_timing.ApplyingState)
+            return;
+
+        if (args.Container.ID != ent.Comp.ContainerId)
+            return;
+
+        if (TryComp(args.Entity, out LockableEquipmentComponent? device))
+            ResetAppearance(ent.Owner, device);
+    }
+
+    private void OnUseHeldKeyVerb(Entity<EquipmentContainerComponent> ent, ref EquipmentContainerUseHeldKeyVerbEvent args)
+    {
+        if (args.User == EntityUid.Invalid || Deleted(args.User))
+            return;
+
+        TryUseHeldKey(ent, args.User);
+    }
+
+    private void OnBreakWithHeldToolVerb(Entity<EquipmentContainerComponent> ent, ref EquipmentContainerBreakWithHeldToolVerbEvent args)
+    {
+        if (args.User == EntityUid.Invalid || Deleted(args.User))
+            return;
+
+        TryBreakWithHeldTool(ent, args.User);
+    }
+
+    private void OnRemoveVerb(Entity<EquipmentContainerComponent> ent, ref EquipmentContainerRemoveVerbEvent args)
+    {
+        if (args.User == EntityUid.Invalid || Deleted(args.User))
+            return;
+
+        TryRemove(ent.Owner, args.User);
+    }
+
+    /// <summary>
+    /// Attempts to remove the currently installed device from the target.
+    /// </summary>
+    public void TryRemove(EntityUid target, EntityUid user)
+    {
+        if (user == EntityUid.Invalid || Deleted(user))
+            return;
+
+        if (!TryComp(target, out EquipmentContainerComponent? comp))
+            return;
+
+        EnsureComp<DoAfterComponent>(user);
+
+        var container = _container.EnsureContainer<ContainerSlot>(target, comp.ContainerId);
+        if (!CanRemove(target, user, container))
+            return;
+
+        var doAfter = new DoAfterArgs(
+            EntityManager,
+            user,
+            comp.DetachDoAfter,
+            new EquipmentDoAfterEvent(EquipmentActionType.Detach),
+            target,
+            target: target)
+        {
+            BreakOnMove = true,
+            BreakOnDamage = true,
+            NeedHand = true,
+            BreakOnHandChange = true,
+            DuplicateCondition = DuplicateConditions.SameTarget | DuplicateConditions.SameEvent
+        };
+
+        _doAfter.TryStartDoAfter(doAfter);
+    }
+
+    /// <summary>
+    /// Attempts to start the attach flow for a device held by the user.
+    /// </summary>
+    public bool TryAttachDevice(
+        Entity<EquipmentContainerComponent> ent,
+        EntityUid user,
+        EntityUid deviceUid,
+        LockableEquipmentComponent device,
+        BaseContainer? container = null)
+    {
+        if (user == EntityUid.Invalid || Deleted(user))
+            return false;
+
+        container ??= _container.EnsureContainer<ContainerSlot>(ent.Owner, ent.Comp.ContainerId);
+
+        var accessAllowed = CanAccess(ent.Owner, device.Layer, device);
+        if (!accessAllowed)
+        {
+            _popup.PopupClient(
+                Loc.GetString("lockable-equipment-blocked"),
+                user);
+            return false;
+        }
+
+        if (FindDevice(container) != null)
+        {
+            _popup.PopupClient(
+                Loc.GetString("lockable-equipment-already"),
+                user);
+            return false;
+        }
+
+        EnsureComp<DoAfterComponent>(user);
+
+        var doAfter = new DoAfterArgs(
+            EntityManager,
+            user,
+            ent.Comp.AttachDoAfter,
+            new EquipmentDoAfterEvent(EquipmentActionType.Attach),
+            ent.Owner,
+            target: ent.Owner,
+            used: deviceUid)
+        {
+            BreakOnMove = true,
+            BreakOnDamage = true,
+            NeedHand = true,
+            BreakOnHandChange = true,
+            DuplicateCondition = DuplicateConditions.SameTarget | DuplicateConditions.SameEvent
+        };
+
+        return _doAfter.TryStartDoAfter(doAfter);
+    }
+
     private void UpdateAppearance(EntityUid uid, LockableEquipmentComponent device)
     {
         if (!TryComp(uid, out AppearanceComponent? appearance))
             return;
 
-        var visualData = CreateVisualData(device, visible: true);
-
+        var visualData = CreateVisualData(device, true);
         _appearance.SetData(uid, EquipmentVisuals.VisualData, visualData, appearance);
     }
 
@@ -245,11 +361,12 @@ public sealed class EquipmentContainerSystem : EntitySystem
 
         if (previousDevice != null)
         {
-            visualData = CreateVisualData(previousDevice, visible: false);
+            visualData = CreateVisualData(previousDevice, false);
         }
         else
         {
             _appearance.TryGetData(uid, EquipmentVisuals.VisualData, out visualData, appearance);
+
             if (visualData != null)
                 visualData = new EquipmentVisualData(false, visualData.Layer, visualData.RsiPath, visualData.State);
         }
@@ -288,68 +405,6 @@ public sealed class EquipmentContainerSystem : EntitySystem
         return _layerAccess.IsLayerAccessible(owner, layer, device);
     }
 
-    private void OnContainerInserted(Entity<EquipmentContainerComponent> ent, ref EntInsertedIntoContainerMessage args)
-    {
-        if (_timing.ApplyingState)
-            return;
-
-        if (args.Container.ID != ent.Comp.ContainerId)
-            return;
-
-        if (TryComp(args.Entity, out LockableEquipmentComponent? device))
-            UpdateAppearance(ent.Owner, device);
-    }
-
-    private void OnContainerRemoved(Entity<EquipmentContainerComponent> ent, ref EntRemovedFromContainerMessage args)
-    {
-        if (_timing.ApplyingState)
-            return;
-
-        if (args.Container.ID != ent.Comp.ContainerId)
-            return;
-
-        if (TryComp(args.Entity, out LockableEquipmentComponent? device))
-            ResetAppearance(ent.Owner, device);
-    }
-
-    private static EquipmentVisualData CreateVisualData(LockableEquipmentComponent device, bool visible)
-    {
-        return new EquipmentVisualData(
-            visible,
-            device.Layer,
-            device.RsiPath,
-            device.SpriteState);
-    }
-
-    private bool CanAttachDevice(EntityUid owner, LockableEquipmentComponent device, BaseContainer container)
-    {
-        return CanAccess(owner, device.Layer, device) && FindDevice(container) == null;
-    }
-
-    private void OnUseHeldKeyVerb(Entity<EquipmentContainerComponent> ent, ref EquipmentContainerUseHeldKeyVerbEvent args)
-    {
-        if (args.User == EntityUid.Invalid || Deleted(args.User))
-            return;
-
-        TryUseHeldKey(ent, args.User);
-    }
-
-    private void OnBreakWithHeldToolVerb(Entity<EquipmentContainerComponent> ent, ref EquipmentContainerBreakWithHeldToolVerbEvent args)
-    {
-        if (args.User == EntityUid.Invalid || Deleted(args.User))
-            return;
-
-        TryBreakWithHeldTool(ent, args.User);
-    }
-
-    private void OnRemoveVerb(Entity<EquipmentContainerComponent> ent, ref EquipmentContainerRemoveVerbEvent args)
-    {
-        if (args.User == EntityUid.Invalid || Deleted(args.User))
-            return;
-
-        TryRemove(ent.Owner, args.User);
-    }
-
     private bool TryUseHeldKey(Entity<EquipmentContainerComponent> ent, EntityUid user)
     {
         if (!TryGetEquipment(ent.Owner, ent.Comp, out var device))
@@ -364,6 +419,28 @@ public sealed class EquipmentContainerSystem : EntitySystem
                 continue;
 
             return _lockable.TryUseKey(device, held.Value, user);
+        }
+
+        return false;
+    }
+
+    private bool TryBreakWithHeldTool(Entity<EquipmentContainerComponent> ent, EntityUid user)
+    {
+        if (!TryGetEquipment(ent.Owner, ent.Comp, out var device))
+            return false;
+
+        if (!TryComp(device, out LockableEquipmentComponent? comp))
+            return false;
+
+        foreach (var hand in _hands.EnumerateHands(user))
+        {
+            if (!_hands.TryGetHeldItem(user, hand, out var held))
+                continue;
+
+            if (!_lockable.CanBreakWithTool(device, held.Value, comp))
+                continue;
+
+            return _lockable.TryStartBreakDoAfter(device, held.Value, user, ent);
         }
 
         return false;
@@ -427,26 +504,21 @@ public sealed class EquipmentContainerSystem : EntitySystem
         return false;
     }
 
-    private bool TryBreakWithHeldTool(Entity<EquipmentContainerComponent> ent, EntityUid user)
+    private string? GetBreakVerbText(string name, LockableEquipmentComponent.BreakMode mode)
     {
-        if (!TryGetEquipment(ent.Owner, ent.Comp, out var device))
-            return false;
-
-        if (!TryComp(device, out LockableEquipmentComponent? comp))
-            return false;
-
-        foreach (var hand in _hands.EnumerateHands(user))
+        return mode switch
         {
-            if (!_hands.TryGetHeldItem(user, hand, out var held))
-                continue;
+            LockableEquipmentComponent.BreakMode.ForceOpen =>
+                Loc.GetString("lockable-equipment-verb-force-open", ("name", name)),
 
-            if (!_lockable.CanBreakWithTool(device, held.Value, comp))
-                continue;
+            LockableEquipmentComponent.BreakMode.Breakable =>
+                Loc.GetString("lockable-equipment-verb-break", ("name", name)),
 
-            return _lockable.TryStartBreakDoAfter(device, held.Value, user, ent);
-        }
+            LockableEquipmentComponent.BreakMode.Destroyable =>
+                Loc.GetString("lockable-equipment-verb-destroy", ("name", name)),
 
-        return false;
+            _ => null
+        };
     }
 
     private void PopupEquipmentEquipped(EntityUid device, EntityUid user)
@@ -500,5 +572,14 @@ public sealed class EquipmentContainerSystem : EntitySystem
         }
 
         return true;
+    }
+
+    private static EquipmentVisualData CreateVisualData(LockableEquipmentComponent device, bool visible)
+    {
+        return new EquipmentVisualData(
+            visible,
+            device.Layer,
+            device.RsiPath,
+            device.SpriteState);
     }
 }
